backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
google-genai>=0.6.2
firebase-admin>=6.5.0
python-jose[cryptography]>=3.3.0
pydantic>=2.6.0
httpx

# AI/ML Dependencies
google-genai
google-cloud-aiplatform>=0.26.0
